fix bug security

+++ OurResumeIR.Application/Services/Implementation/SkillLevelService.cs
@@ -1,10 +1,11 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Mvc.Rendering;
-using OurResumeIR.Application.ViewModels.Experience;
 using OurResumeIR.Application.ViewModels.ExpertiseLayers;
 using OurResumeIR.Application.ViewModels.MySkills;
+using OurResumeIR.Application.ViewModels.SkillLevel;
 using OurResumeIR.Domain.Interfaces;
 using OurResumeIR.Domain.Models;
+using System.Xml.Linq;
 
 namespace OurResumeIR.Application.Services.Interfaces;
 
@@ -27,6 +28,11 @@ public class SkillLevelService(IUnitOfWork unitOfWork, IMapper mapper) : ISkillL
 
     }
 
+    public async Task<List<SkillLevelVM>> GetAll(string userId)
+    {
+        return mapper.Map <List<SkillLevelVM>>( await unitOfWork.SkillLevelRepository.GetAllSkillLevelAsync(userId));
+    }
+
     public async Task<SkillLevelVM> GetById(int Id)
     {
         var rep_SkillLevel = unitOfWork.SkillLevelRepository;
@@ -36,20 +42,21 @@ public class SkillLevelService(IUnitOfWork unitOfWork, IMapper mapper) : ISkillL
     public async Task<bool> Update(UpdateSkillLevelVM model)
     {
         var rep_SkillLevel = unitOfWork.SkillLevelRepository;
-        var newModel = new SkillLevel()
-        {
-            Id = model.Id,
-            Name = model.Name,
-            Percentage = model.Percentage
-        };
-        bool status = await rep_SkillLevel.UpdateSkillLevelLevel(newModel);
+     
+        var skillLevel = await rep_SkillLevel.GetSkillLevelById(model.Id);
+        if (skillLevel == null | !skillLevel.UserId.Equals(model.UserId))
+            return false;
+        //bool status = await rep_SkillLevel.UpdateSkillLevelLevel(newModel);
+        skillLevel.Name = model.Name;
+        skillLevel.Percentage = model.Percentage;
         await rep_SkillLevel.SaveChanges();
 
-        return status;
+        return true;
     }
 
     public async Task<bool> Create(CreateSkillLevelVM model)
     {
+        
         var rep_SkillLevel = unitOfWork.SkillLevelRepository;
         var newModel = mapper.Map<SkillLevel>(model);
         int id = await rep_SkillLevel.CreateSkillLevelLevel(newModel);
@@ -61,9 +68,13 @@ public class SkillLevelService(IUnitOfWork unitOfWork, IMapper mapper) : ISkillL
 
     }
 
-    public async Task<bool> Delete(int Id)
+
+    public async Task<bool> Delete(int Id,string userId)
     {
         var rep_SkillLevel = unitOfWork.SkillLevelRepository;
+        var skillLevel =await rep_SkillLevel.GetSkillLevelById(Id);
+        if (skillLevel == null | !skillLevel.UserId.Equals(userId))
+            return false;
         var status = await rep_SkillLevel.DeleteSkillLevelLevel(Id);
         await rep_SkillLevel.SaveChanges();
         return status;
